shows events data in events master
