+++ Dinder/Base.lproj/Main.storyboard
@@ -255,8 +255,11 @@
                                             </constraints>
                                         </tableViewCellContentView>
                                         <color key="backgroundColor" red="1" green="0.91896892969999999" blue="0.61512825429999995" alpha="1" colorSpace="custom" customColorSpace="displayP3"/>
+                                        <connections>
+                                            <outlet property="activeEventController" destination="zte-r9-Krd" id="Px4-Wj-tdy"/>
+                                        </connections>
                                     </tableViewCell>
-                                    <tableViewCell clipsSubviews="YES" contentMode="scaleToFill" preservesSuperviewLayoutMargins="YES" selectionStyle="default" indentationWidth="10" reuseIdentifier="myInterestedEventCell" rowHeight="200" id="esb-Pd-Hoy" userLabel="MyInterestedEventCell" customClass="MyActiveEventCell" customModule="Dinder" customModuleProvider="target">
+                                    <tableViewCell clipsSubviews="YES" contentMode="scaleToFill" preservesSuperviewLayoutMargins="YES" selectionStyle="default" indentationWidth="10" reuseIdentifier="myInterestedEventCell" rowHeight="200" id="esb-Pd-Hoy" userLabel="MyInterestedEventCell" customClass="MyInterestedEventCell" customModule="Dinder" customModuleProvider="target">
                                         <rect key="frame" x="0.0" y="272" width="375" height="200"/>
                                         <autoresizingMask key="autoresizingMask"/>
                                         <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" preservesSuperviewLayoutMargins="YES" insetsLayoutMarginsFromSafeArea="NO" tableViewCell="esb-Pd-Hoy" id="K0l-nT-A5j">
@@ -276,6 +279,9 @@
                                             </constraints>
                                         </tableViewCellContentView>
                                         <color key="backgroundColor" red="1" green="0.91896892969999999" blue="0.61512825429999995" alpha="1" colorSpace="custom" customColorSpace="displayP3"/>
+                                        <connections>
+                                            <outlet property="interestedEventController" destination="U3v-7r-PD0" id="bmN-bi-iuh"/>
+                                        </connections>
                                     </tableViewCell>
                                 </prototypes>
                             </tableView>
