+++ Dinder/Events Display/MyActiveEventView.swift
@@ -3,6 +3,16 @@
 import UIKit
 
 class MyActiveEventView: UIView {
+  
+  
+  @IBOutlet weak var eventImage: UIImageView!
+  @IBOutlet weak var eventTitle: UILabel!
+  @IBOutlet weak var eventDate: UILabel!
+  
+  @IBOutlet weak var interestedUsersCollectionView: UICollectionView!
+  
+  
+  
   @IBOutlet var contentView: UIView!
   
   override init(frame: CGRect) {
@@ -20,6 +30,20 @@ class MyActiveEventView: UIView {
     addSubview(contentView)
     contentView.frame = self.bounds
     contentView.autoresizingMask = [.flexibleWidth , .flexibleHeight]
+  }
+  
+  func setUp(event: Event) {
+    //TODO check if images > 0
+    
+    event.images[0].getImage(imageView: eventImage)
+    eventTitle.text = event.title
+    let gbLocale = Locale(identifier: "en_GB")
+    let customFormat = "dMMMMhh"
+    let ukFormat = DateFormatter.dateFormat(fromTemplate: customFormat, options: 0, locale: gbLocale)
+    let formatter = DateFormatter()
+    formatter.dateFormat = ukFormat
+    eventDate.text = formatter.string(from: event.startDateTime)
+    
   }
 
 }
